+++ packages/mobile-visualization/CHANGELOG.md
@@ -8,6 +8,13 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.4.0-beta.7 (12/2/2025 PST)
+
+#### 🐞 Fixes
+
+- Fix babel build issue with transitions. [[#201](https://github.com/coinbase/cds/pull/201)]
+- Improve opacity customization for ReferenceLine. [[#201](https://github.com/coinbase/cds/pull/201)]
+
 ## 3.4.0-beta.6 (10/16/2025 PST)
 
 #### 🚀 Updates